[VIGT-3] - добавила NavController
