+++ VIGT App/VIGT App/Modules/Main/MainViewController.swift
@@ -1,0 +1,16 @@
+//
+//  MainViewController.swift
+//  VIGT App
+//
+//  Created by Наталья Коновалова on 24.11.2024.
+//
+
+import UIKit
+
+final class MainViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .yellow
+    }
+}
