+++ VIGT App/VIGT App/Modules/Registr/RegistrationViewController.swift
@@ -22,12 +22,12 @@ final class RegistrationViewController: UIViewController {
     
     let passwordTextField: CustomTextField = {
         let textField = CustomTextField()
+        textField.font = .systemFont(ofSize: 16)
         textField.placeholder = "Пароль"
         textField.backgroundColor = .init(red: 1, green: 1, blue: 1, alpha: 0.5)
         textField.layer.cornerRadius = 8
         textField.tintColor = .darkGray
         textField.textPadding = UIEdgeInsets(top: 0, left: 10, bottom: 0, right: 0)
-        textField.isSecureTextEntry = true
         textField.translatesAutoresizingMaskIntoConstraints = false
         return textField
     }()
@@ -75,23 +75,24 @@ extension RegistrationViewController {
     private func setupConstraints() {
         NSLayoutConstraint.activate([
             emailTextField.centerXAnchor.constraint(equalTo: view.centerXAnchor),
-            emailTextField.centerYAnchor.constraint(equalTo: view.centerYAnchor, constant: -32),
-            emailTextField.widthAnchor.constraint(equalToConstant: 180),
-            emailTextField.heightAnchor.constraint(equalToConstant: 32)
+            emailTextField.centerYAnchor.constraint(equalTo: view.centerYAnchor, constant: -.spacing32),
+            emailTextField.widthAnchor.constraint(equalToConstant: .spacing184),
+            emailTextField.heightAnchor.constraint(equalToConstant: .spacing32)
         ])
         
         NSLayoutConstraint.activate([
             passwordTextField.centerXAnchor.constraint(equalTo: view.centerXAnchor),
-            passwordTextField.centerYAnchor.constraint(equalTo: view.centerYAnchor, constant: 32),
-            passwordTextField.widthAnchor.constraint(equalToConstant: 180),
-            passwordTextField.heightAnchor.constraint(equalToConstant: 32)
+            passwordTextField.centerYAnchor.constraint(equalTo: view.centerYAnchor, constant: .spacing32),
+            passwordTextField.widthAnchor.constraint(equalToConstant: .spacing184),
+            passwordTextField.heightAnchor.constraint(equalToConstant: .spacing32)
         ])
         
         NSLayoutConstraint.activate([
             registerButton.centerXAnchor.constraint(equalTo: view.centerXAnchor),
-            registerButton.topAnchor.constraint(equalTo: passwordTextField.bottomAnchor, constant: 8),
-            registerButton.heightAnchor.constraint(equalToConstant: 32)
+            registerButton.topAnchor.constraint(equalTo: passwordTextField.bottomAnchor, constant: .spacing8),
+            registerButton.heightAnchor.constraint(equalToConstant: .spacing32)
         ])
     }
 }
+
